Changed email analytics fetch-latest to run on the JobQueue
The five-minute analytics fetch moves off its bree worker thread onto
the JobQueue. The old worker could only post a message that the job
manager relayed as StartEmailAnalyticsJobEvent to reach the service on
the main thread; the handler now calls the analytics service directly,
so that event, its subscription, and the worker file are deleted. The
domain jobs module keeps the scheduling gate (analytics enabled, sent
emails exist) and mega still re-triggers scheduling on first send.
