+++ ghost/core/core/boot.js
@@ -450,8 +450,8 @@ async function initBackgroundServices({config}) {
     await activitypub.init();
     // Load email analytics recurring jobs
     if (config.get('backgroundJobs:emailAnalytics')) {
-        const emailAnalyticsJobs = require('./server/services/email-analytics/jobs');
-        await emailAnalyticsJobs.scheduleRecurringJobs();
+        const emailAnalytics = require('./server/services/email-analytics');
+        await emailAnalytics.scheduleRecurringJobs();
     }
 
     const updateCheck = require('./server/services/update-check');
